+++ MEDIFACILAPP/app/components/Account/pharmacy/PharmacyRegisterForm.js
@@ -1,13 +1,12 @@
-import React, { useState } from "react"
-import { StyleSheet, View, Text} from 'react-native' 
-import { Input, Icon, Button } from 'react-native-elements'
-import {validateEmail} from '../../utils/Validation'
-import firebase from "firebase"
-import {useNavigation} from "@react-navigation/native"
+import React, {useState} from 'react'
+import {StyleSheet, View, Text} from 'react-native'
+import {Input, Icon, Button} from 'react-native-elements'
+import firebase from 'firebase'
+import {useNavigation} from '@react-navigation/native'
+import { validateEmail } from '../../../utils/Validation'
 
 
-
-export default function UserRegisterForm(props){
+export default function PharmacyRegisterForm(props){
     const {toastRef} = (props) 
     const [showPassword, setShowPassword] = useState(false)
     const [showRepeatPassword, setShowRepeatPassword] = useState(false)
@@ -16,7 +15,7 @@ export default function UserRegisterForm(props){
     const navigation = useNavigation()
 
     const onSubmit = () => {
-       if(formData.email.length===0||formData.password.length===0||formData.repeatPassword.length===0){
+        if(formData.rfc.length===0||formData.email.length===0||formData.password.length===0||formData.repeatPassword.length===0){
             toastRef.current.show({
                 type: 'error',
                 position: 'top',
@@ -24,15 +23,24 @@ export default function UserRegisterForm(props){
                 text2: 'Todos los campos son requeridos',
                 visibilityTime: 30000
             });
-       } else if (!validateEmail(formData.email)){
+        } else if (formData.rfc.length < 13){
+        console.log('El rfc debe tener 10 caracteres')
+            toastRef.current.show({
+                type: 'error',
+                position: 'top',
+                text1: 'Empty',
+                text2: 'El RFC debe tener mínimo 13 caracteres',
+                visibilityTime: 30000
+            }); 
+        } else if (!validateEmail(formData.email)){
             toastRef.current.show({
                 type: 'error',
                 position: 'top',
                 text1: 'Empty',
                 text2: 'El email no es correcto',
                 visibilityTime: 30000
-            });  
-       } else if (formData.password !== formData.repeatPassword){
+            }); 
+        } else if (formData.password !== formData.repeatPassword){
             toastRef.current.show({
                 type: 'error',
                 position: 'top',
@@ -40,8 +48,8 @@ export default function UserRegisterForm(props){
                 text2: 'Las contraseñas deben ser identicas',
                 visibilityTime: 30000
             });
-       } else if (formData.password.length < 6){
-           toastRef.current.show({
+        } else if (formData.password.length < 6){
+            toastRef.current.show({
                 type: 'error',
                 position: 'top',
                 text1: 'Empty',
@@ -52,7 +60,7 @@ export default function UserRegisterForm(props){
             firebase.auth()
             .createUserWithEmailAndPassword(formData.email, formData.password)
             .then(() =>{
-                navigation.navigate('account')
+                navigation.navigate('pharmacyLogged')
             })
             .catch(()=>{
                 toastRef.current.show({
@@ -72,6 +80,12 @@ export default function UserRegisterForm(props){
 
     return(
         <View style={styles.formContainer}>
+            <Input
+            placeholder="RFC"
+            containerStyle={styles.inputForm}
+            onChange={(e)=>onChange(e, 'rfc')}
+            rightIcon={<Icon type='material-community' name= 'text-box-outline' iconStyle={styles.iconRight}/>}
+            />
             <Input
             placeholder="Correo electronico"
             containerStyle={styles.inputForm}
@@ -105,7 +119,7 @@ export default function UserRegisterForm(props){
                 />}
             />
             <Button
-            title='Únete'
+            title='Resgistrarse'
             containerStyle={styles.btnContainerRegister}
             buttonStyle={styles.btnRegister}
             onPress={onSubmit}
@@ -116,6 +130,7 @@ export default function UserRegisterForm(props){
 
 function defaultFormValues(){
     return{
+        rfc: '',
         email: '',
         password: '',
         repeatPassword: ''
