+++ MEDIFACILAPP/app/components/Account/user/UserRegisterForm.js
@@ -1,12 +1,13 @@
-import React, {useState} from 'react'
-import {StyleSheet, View, Text} from 'react-native'
-import {Input, Icon, Button} from 'react-native-elements'
-import firebase from 'firebase'
-import {useNavigation} from '@react-navigation/native'
-import { validateEmail } from '../../utils/Validation'
+import React, { useState } from "react"
+import { StyleSheet, View, Text} from 'react-native' 
+import { Input, Icon, Button } from 'react-native-elements'
+import {validateEmail} from '../../../utils/Validation'
+import firebase from "firebase"
+import {useNavigation} from "@react-navigation/native"
 
 
-export default function PharmacyRegisterForm(props){
+
+export default function UserRegisterForm(props){
     const {toastRef} = (props) 
     const [showPassword, setShowPassword] = useState(false)
     const [showRepeatPassword, setShowRepeatPassword] = useState(false)
@@ -15,7 +16,7 @@ export default function PharmacyRegisterForm(props){
     const navigation = useNavigation()
 
     const onSubmit = () => {
-        if(formData.rfc.length===0||formData.email.length===0||formData.password.length===0||formData.repeatPassword.length===0){
+       if(formData.email.length===0||formData.password.length===0||formData.repeatPassword.length===0){
             toastRef.current.show({
                 type: 'error',
                 position: 'top',
@@ -23,24 +24,15 @@ export default function PharmacyRegisterForm(props){
                 text2: 'Todos los campos son requeridos',
                 visibilityTime: 30000
             });
-        } else if (formData.rfc.length < 13){
-        console.log('El rfc debe tener 10 caracteres')
-            toastRef.current.show({
-                type: 'error',
-                position: 'top',
-                text1: 'Empty',
-                text2: 'El RFC debe tener mínimo 13 caracteres',
-                visibilityTime: 30000
-            }); 
-        } else if (!validateEmail(formData.email)){
+       } else if (!validateEmail(formData.email)){
             toastRef.current.show({
                 type: 'error',
                 position: 'top',
                 text1: 'Empty',
                 text2: 'El email no es correcto',
                 visibilityTime: 30000
-            }); 
-        } else if (formData.password !== formData.repeatPassword){
+            });  
+       } else if (formData.password !== formData.repeatPassword){
             toastRef.current.show({
                 type: 'error',
                 position: 'top',
@@ -48,8 +40,8 @@ export default function PharmacyRegisterForm(props){
                 text2: 'Las contraseñas deben ser identicas',
                 visibilityTime: 30000
             });
-        } else if (formData.password.length < 6){
-            toastRef.current.show({
+       } else if (formData.password.length < 6){
+           toastRef.current.show({
                 type: 'error',
                 position: 'top',
                 text1: 'Empty',
@@ -60,7 +52,7 @@ export default function PharmacyRegisterForm(props){
             firebase.auth()
             .createUserWithEmailAndPassword(formData.email, formData.password)
             .then(() =>{
-                navigation.navigate('pharmacyLogged')
+                navigation.navigate('account')
             })
             .catch(()=>{
                 toastRef.current.show({
@@ -80,12 +72,6 @@ export default function PharmacyRegisterForm(props){
 
     return(
         <View style={styles.formContainer}>
-            <Input
-            placeholder="RFC"
-            containerStyle={styles.inputForm}
-            onChange={(e)=>onChange(e, 'rfc')}
-            rightIcon={<Icon type='material-community' name= 'text-box-outline' iconStyle={styles.iconRight}/>}
-            />
             <Input
             placeholder="Correo electronico"
             containerStyle={styles.inputForm}
@@ -119,7 +105,7 @@ export default function PharmacyRegisterForm(props){
                 />}
             />
             <Button
-            title='Resgistrarse'
+            title='Únete'
             containerStyle={styles.btnContainerRegister}
             buttonStyle={styles.btnRegister}
             onPress={onSubmit}
@@ -130,7 +116,6 @@ export default function PharmacyRegisterForm(props){
 
 function defaultFormValues(){
     return{
-        rfc: '',
         email: '',
         password: '',
         repeatPassword: ''
